requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
demoji==0.4.0
elastic-enterprise-search==7.11.0
elastic-transport==7.11.0
emoji==1.2.0
expertai-nlapi==2.1.3
Flask @ file:///home/ktietz/src/ci/flask_1611932660458/work
gunicorn==20.0.4
idna==2.10
itsdangerous @ file:///home/ktietz/src/ci/itsdangerous_1611932585308/work
Jinja2 @ file:///tmp/build/80754af9/jinja2_1612213139570/work
MarkupSafe @ file:///opt/concourse/worker/volumes/live/cb778296-98db-45ad-411e-6f726e102dc3/volume/markupsafe_1594371638608/work
PyJWT==2.0.1
python-dateutil==2.8.1
requests==2.25.1
six==1.15.0
soupsieve==2.2
urllib3==1.26.3
Werkzeug @ file:///home/ktietz/src/ci/werkzeug_1611932622770/work
